Add build cookbook; Update README
Add the initial workings of a build cookbook. The focus now is on the
lint/unit phases as we iterate. Once a certain quality has been met we
will start publishing to external sources for public consumption.

This change also includes updates to the README to add a few more
details.