Add changed_files support + testing
This adds `changed_files` support along with updates and improvements to
the functional and unit test suites.

We are accomplishing functional tests for our `changed_files` method by
including a git-bundle that we can explode and then inspect. There are
instructions in the included README for how this was set up as well as
how to amend it.