A missing change sha is "" not nil; Use origin
The `changed_files` method was making a lot of incorrect assumptions.

1. It incorrectly assumed that a missing change sha was nil rather than
an empty string.
2. It assumed that the pipeline_branch be checked out locally. That is
not the case. It will need to do the merge-base check against the remote
branch.
3. We did not account for the 'pseudo-merge' that was occuring. What the
delivery-cli is doing is merging the pipeline branch in its local copy of
the pipleine branch. That means we'll need to perform the merge-base against the
remote pipeline branch, not the local copy.

The shell_outs in SCM::Git were also supressing errors. To avoid this we
are now using `shell_out!`.