Added delivery_chef_cookbook LWRP
As a Delivery user, I would like to have a `delivery_chef_cookbook`
resource that I can use to upload a cookbook to a Chef Server.

When you create a new `DeliverySugar::ChefServer` object, we now
remember the path to the `knife.rb` that you provided. That allows us to
execute `knife` commands against that Chef Server.