Replace `cookbook_name` attribute for `cookbook_to_upload`
As a user that uses delivery-sugar, I want to be able to use the
defaults of `delivery_chef_cookbook` with out failing so I don't
have to specify every single attribute to the resource.

**Problem Description**

We used an attribute name that is currently used by Chef::Resource
(https://github.com/chef/chef/blob/master/lib/chef/resource.rb#L1492-L1493)
called `cookbook_name` so it gets overriden by the actual name of
the cookbook that is running on, in this case the `build-cookbook`
within delivery.

**Proposal Solution**

Change the attribute name to `cookbook_to_upload`