Consume delivery-cli attribute workspace_path
This commit is going to consume the new attribute `workspace_path`
coming from the `delivery-cli`. The change is under review here:
https://delivery.shd.chef.co/e/Chef/#/organizations/Chef_Delivery/projects/delivery-cli/changes/73157eba-57fe-4877-aea0-0c44b1edf6ce/review

This change is secure enough to keep working with old versions of the
CLI but ready to pick the new changes coming. With this change we are
going to be in consintent with the workspace that the `delivery_build`
cookbook provides.

Closes https://github.com/chef-cookbooks/delivery-sugar/issues/2