Handling secrets at the organization level
Currently the way we load secrets is by project. We have to have a item
called `ent-org-proj` to load the secrets of the project from there. If
the item does not exist we will fail.

This change allows you to load **organization** level secrets so that
your projects can load shared secrets within your organization. It will
expect to have a item called `ent-org`.

Additionally if the project is trying to pull down the secrets of a
project that does not exist, delivery-sugar automatically will try to
load the organization level secrets.

Closes https://github.com/chef-cookbooks/delivery-sugar/issues/5