Remove gem dependency from metadata
This change is removing the `gem 'chef-vault'`from the `metadata.rb`
since it is breaking air-gapped environments because is trying to reach
out to rubygems.org

The right solution is to add an option to the `automate-ctl install-runner`
command that sets up the `Chef::Config[:rubygems_url]` to the users rubygem
internal mirror and then put back the `gem` dependency in the
`metadata`. This will solve all the gem problems on air-gapped
environments moving forward.

Signed-off-by: Salim Afiune <afiune@chef.io>