adding delivery_terraform resource
This change adds a `delivery_terraform` resource to the delivery-sugar
cookbook in order to support use cases where ephemeral nodes are needed
for testing during the workflow stages.  Probably the best opportunity
for use would be in the Acceptance stage.

Along with the libraries implementing the resource, there is a full
fledged example of a build cookbook utilizing the resource under
`examples/` and a README_TERRAFORM.md documenting the resource and
its usage with more examples.

Signed-off-by: Jeremy J. Miller <jm@chef.io>

converting to custom resource

Signed-off-by: Jeremy J. Miller <jm@chef.io>

removed uneeded test

Signed-off-by: Jeremy J. Miller <jm@chef.io>

bring custom resource back

Signed-off-by: Jeremy J. Miller <jm@chef.io>