This change:
- removes the `repo_path` delivery_terraform property and replaces it with
  delivery_workspace_repo in all locations within the resource
- removes `plan_dir` default value within delivery_terraform and makes it
  a required property
- adds a default spec for ensuring a downstream cookbook can compile
  `delivery-sugar` outside the delivery cli context where
  node['delivery']... is not defined
- bumps the patch version

In support of the above, the below was removed from the delivery_terraform
resource:
```
extend DeliverySugar::DSL

property :plan_dir, String, default:
"#{delivery_workspace_repo}/files/default/terraform"
property :repo_path, String, default: delivery_workspace_repo
```

The `extend` and invocation of the `delivery_workspace_repo` method to
assign the default properities caused a side effect of calling
`Delivery::Change.new` which required `node['delivery']..` to be set.

This would fail cookbook compile with `undefined method []' for
nil:NilClass` if delivery-sugar was included in a run_list outside
delivery-cli (Build Node/Runner) context.

That was the case for the the last commit of this cookbook which failed
acceptance here:
https://automate.chef.co/e/chef/#/organizations/Delivery-Build-Cookbooks/projects/delivery-sugar/changes/e3960d3e-56b9-41a9-9695-0d71bd994bc2/status/acceptance

The above is a functional test failure during the Test Kitchen converge of
`test-build-cookbook::default` fixture cookbook recipe.  TK converges
the EC2 node with a declared dependency on `deliver-sugar` in the
TK node's expanded run_list.  This forces a compile of `delivery-sugar`
without `node['delivery']..` set because it's outside the context of
delivery-cli resulting in a compile time failure.

This change is targeted at correcting this problem scenario so that
the delivery_terraform resource no longer forces a compile time
`Delivery::Change.new` call, allowing it to compile succesfully, even
if the `node['delivery']..` attributes do not exist.

Signed-off-by: Jeremy J. Miller <jm@chef.io>